+++ common.h
@@ -1,0 +1,6 @@
+#ifndef __COMMON_H
+#define __COMMON_H
+
+typedef unsigned int size_t;
+
+#endif
