+++ frequencies.cpp
@@ -1,0 +1,16 @@
+#include "frequencies.h"
+
+size_t operator "" _Hz(unsigned long long value)
+{
+    return value;
+}
+
+size_t operator "" _KHz(unsigned long long value)
+{
+    return 1000 * value;
+}
+
+size_t operator "" _MHz(unsigned long long value)
+{
+    return 1000_KHz * value;
+}
